final commit

+++ controllers/blogs.js
@@ -1,15 +1,6 @@
-const jwt = require('jsonwebtoken')
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
 const User = require('../models/user')
-
-const getTokenFrom = request => {
-    const authorization = request.get('authorization')
-    if (authorization && authorization.toLowerCase().startsWith('bearer ')) {
-        return authorization.substring(7)
-    }
-    return null
-}
 
 blogsRouter.get('/', async (request, response) => {
     const blogs = await Blog
@@ -20,20 +11,16 @@ blogsRouter.get('/', async (request, response) => {
 
 blogsRouter.post('/', async (request, response) => {
     const body = request.body
-    const token = getTokenFrom(request)
-    const decodedToken = jwt.verify(token, process.env.SECRET)
-    if (!token || !decodedToken.id) {
-        return response.status(401).json({ error: 'token missing or invalid'})
-    }
 
-    const user = await User.findById(decodedToken.id)
+    const userId = request.user
+    const user = await User.findById(userId)
 
     let blog = new Blog({
         title: body.title,
         author: body.author,
         url: body.url,
         likes: body.likes,
-        user: user._id
+        user: userId
     })
 
     if (!blog.likes) {
@@ -50,22 +37,23 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-    try {
-        await Blog.findByIdAndRemove(request.params.id)
-        response.status(204).end()
-    } catch(e) {
-        response.status(404).end()
+    const userId = request.user
+    const blog = await Blog.findById(request.params.id)
+    if (!userId || !blog) { return response.status(404).end() }
+    if (!request.token || blog.user.toString() !== userId) {
+        return response.status(401).json({ error: 'token missing or invalid' })
     }
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+    
 })
 
 blogsRouter.put('/:id', async (request, response) => {
     const blog = new Blog(request.body)
-    try {
-        await Blog.findByIdAndUpdate(request.params.id, blog, { new: true })
-        response.status(204).end()
-    } catch(e) {
-        response.status(404).end()
-    }
+    const re = await Blog.findByIdAndUpdate(request.params.id, blog, { new: true })
+    re
+        ? response.status(204).end()
+        : response.status(404).end()
 })
 
 module.exports = blogsRouter
